Refactored camera to support new camera mechanisms
